Decompose narrative.py into focused submodules
Reduce narrative.py from 2,679 to 708 lines (74% reduction) by extracting
functionality into dedicated modules:

- narrative_schemas.py: 21 Pydantic request/response models
- slot_endpoints.py: 6 slot management API routes
- setup_endpoints.py: 6 wizard setup API routes
- wizard_chat.py: Chat and transition endpoints
- narrative_generation.py: Async generation orchestration

Also:
- Add [global.api] and [api.constraints] config sections to nexus.toml
- Add corresponding Pydantic models to settings_models.py
- Remove 3 obsolete/broken test files that had stale APIs or missing fixtures
